Backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.0.1
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256=5e6329fe3fecedb1d1a02c3ea2172ee0fede6cea6e4aefb6a02d832dba78a310
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fasttext==0.9.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
google_search_results==2.4.2
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.1.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.5
jupyter-tabnine==1.2.3
jupyter_client==8.6.3
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.2
jupyter_nbextensions_configurator==0.6.4
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.14
keras==3.9.2
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.42
language_data==1.3.0
Levenshtein==0.27.1
libclang==18.1.1
lxml==5.3.2
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
mysql-connector-python==9.3.0
namex==0.0.8
narwhals==1.39.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.0
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
openai==1.78.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.6.0
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-json-logger==3.3.0
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
routers==0.10.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikeras==0.13.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-text==2.19.0
termcolor==3.0.1
terminado==0.18.1
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wikipedia==1.4.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
